End-to-End Projects/ThunderCast/requirements.txt
# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning / Forecasting
prophet>=1.1.0

# Database
supabase>=2.0.0
python-dotenv>=1.0.0

# Visualization
plotly>=5.18.0

# Dashboard
streamlit>=1.31.0

# API & Requests
requests>=2.31.0